Finish the warning sweep the build cache had been hiding
The "zero warnings" check after 2.0.3 was run against a partly cached build,
so it was not the clean check I took it for. A full compile with the cache
pointed elsewhere surfaces thirty more, and this clears them:

- Twenty-six locals that exist only to feed a LOG_DBG the release build
  compiles out, now [[maybe_unused]] rather than silenced by flag.
- A third member-initializer order that disagreed with its declarations
  (ParagraphStreamer, the KOReader progress mapper).
- The sleep-screen migration compared an enum against the uint8_t it had just
  been clamped into.
- The font manifest's two range-for loops bind their JsonArray to a named
  local: same lifetime, but GCC 13 stops reading it as a reference into a
  dead proxy.
- WifiSelectionState::PASSWORD_ENTRY is named in the render switch. It is
  unreachable there (render returns above it), and saying so keeps the switch
  exhaustive for every state that is not.

Also: a USB reflash leaves no marker behind, so the new boot diagnostic
reported every development flash as an unexpected shutdown. It now reads
ESP_RST_UNKNOWN as "came back from a flash", the same way HalGPIO already does.

Both environments compile from scratch with no warnings from our own code.
117/117 host tests, cppcheck clean, i18n validation passed, release at 89.8%.

+++ src/JsonSettingsIO.cpp
@@ -233,8 +233,8 @@ bool JsonSettingsIO::loadSettings(CrossPointSettings& s, const char* json, bool*
       clamp(doc["sleepScreen"] | static_cast<uint8_t>(CrossPointSettings::SLEEP_SCREEN_MODE::LIGHT),
             CrossPointSettings::SLEEP_SCREEN_MODE_COUNT,
             static_cast<uint8_t>(CrossPointSettings::SLEEP_SCREEN_MODE::LIGHT));
-  s.sleepScreen = storedSleepScreen == CrossPointSettings::SLEEP_SCREEN_MODE::DARK
-                      ? CrossPointSettings::SLEEP_SCREEN_MODE::LIGHT
+  s.sleepScreen = storedSleepScreen == static_cast<uint8_t>(CrossPointSettings::SLEEP_SCREEN_MODE::DARK)
+                      ? static_cast<uint8_t>(CrossPointSettings::SLEEP_SCREEN_MODE::LIGHT)
                       : storedSleepScreen;
   if (storedSleepScreen == CrossPointSettings::SLEEP_SCREEN_MODE::DARK && needsResave) {
     *needsResave = true;
